Move generic/specific into a "stock" filter package.
