+++ app/src/main/java/eu/darken/sdmse/systemcleaner/core/filter/stock/AnalyticsFilter.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.systemcleaner.core.filter.specific
+package eu.darken.sdmse.systemcleaner.core.filter.stock
 
 import dagger.Binds
 import dagger.Module
@@ -14,49 +14,49 @@ import eu.darken.sdmse.common.ca.CaString
 import eu.darken.sdmse.common.ca.toCaDrawable
 import eu.darken.sdmse.common.ca.toCaString
 import eu.darken.sdmse.common.datastore.value
-import eu.darken.sdmse.common.debug.logging.Logging.Priority.INFO
 import eu.darken.sdmse.common.debug.logging.log
 import eu.darken.sdmse.common.debug.logging.logTag
 import eu.darken.sdmse.common.files.APathLookup
 import eu.darken.sdmse.common.files.segs
-import eu.darken.sdmse.common.root.RootManager
-import eu.darken.sdmse.common.root.canUseRootNow
 import eu.darken.sdmse.systemcleaner.core.BaseSieve
 import eu.darken.sdmse.systemcleaner.core.SystemCleanerSettings
 import eu.darken.sdmse.systemcleaner.core.filter.SystemCleanerFilter
+import java.io.File
 import javax.inject.Inject
 import javax.inject.Provider
 
-class DownloadCacheFilter @Inject constructor(
+class AnalyticsFilter @Inject constructor(
     private val baseSieveFactory: BaseSieve.Factory,
     private val areaManager: DataAreaManager,
 ) : SystemCleanerFilter {
 
-    override suspend fun getIcon(): CaDrawable = R.drawable.ic_android_studio_24.toCaDrawable()
+    override suspend fun getIcon(): CaDrawable = R.drawable.ic_analytics_onsurface.toCaDrawable()
 
-    override suspend fun getLabel(): CaString = R.string.systemcleaner_filter_downloadcache_label.toCaString()
+    override suspend fun getLabel(): CaString = R.string.systemcleaner_filter_analytics_label.toCaString()
 
-    override suspend fun getDescription(): CaString = R.string.systemcleaner_filter_downloadcache_summary.toCaString()
+    override suspend fun getDescription(): CaString = R.string.systemcleaner_filter_analytics_summary.toCaString()
 
     override suspend fun targetAreas(): Set<DataArea.Type> = setOf(
-        DataArea.Type.DOWNLOAD_CACHE,
+        DataArea.Type.SDCARD,
+        DataArea.Type.PUBLIC_DATA,
+        DataArea.Type.PRIVATE_DATA,
     )
 
     private lateinit var sieve: BaseSieve
 
     override suspend fun initialize() {
+        val pathContains = setOf(
+            segs(".bugsense")
+        )
+        val regexes = setOf(
+            Regex("^(?:[\\W\\w]+/\\.(?:bugsense))$".replace("/", "\\" + File.separator))
+        )
+
         val config = BaseSieve.Config(
-            targetTypes = setOf(BaseSieve.TargetType.FILE),
             areaTypes = targetAreas(),
-            exclusions = setOf(
-                BaseSieve.Exclusion(segs("dalvik-cache")),
-                BaseSieve.Exclusion(segs("lost+found")),
-                // Some apps use these logs to determine the type of recovery
-                BaseSieve.Exclusion(segs("recovery", "last_log")),
-                BaseSieve.Exclusion(segs("last_postrecovery")),
-                BaseSieve.Exclusion(segs("last_data_partition_info")),
-                BaseSieve.Exclusion(segs("last_dataresizing")),
-            )
+            targetTypes = setOf(BaseSieve.TargetType.FILE),
+            pathContains = pathContains,
+            regexes = regexes,
         )
         sieve = baseSieveFactory.create(config)
         log(TAG) { "initialized()" }
@@ -72,17 +72,9 @@ class DownloadCacheFilter @Inject constructor(
     @Reusable
     class Factory @Inject constructor(
         private val settings: SystemCleanerSettings,
-        private val filterProvider: Provider<DownloadCacheFilter>,
-        private val rootManager: RootManager,
+        private val filterProvider: Provider<AnalyticsFilter>
     ) : SystemCleanerFilter.Factory {
-
-        override suspend fun isEnabled(): Boolean {
-            val enabled = settings.filterDownloadCacheEnabled.value()
-            val useRoot = rootManager.canUseRootNow()
-            if (enabled && !useRoot) log(TAG, INFO) { "Filter is enabled, but requires root, which is unavailable." }
-            return enabled && useRoot
-        }
-
+        override suspend fun isEnabled(): Boolean = settings.filterAnalyticsEnabled.value()
         override suspend fun create(): SystemCleanerFilter = filterProvider.get()
     }
 
@@ -93,6 +85,6 @@ class DownloadCacheFilter @Inject constructor(
     }
 
     companion object {
-        private val TAG = logTag("SystemCleaner", "Filter", "DownloadCache")
+        private val TAG = logTag("SystemCleaner", "Filter", "AnalyticsFiles")
     }
 }
