+++ app/src/main/java/eu/darken/sdmse/systemcleaner/core/filter/stock/LinuxFilesFilter.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.systemcleaner.core.filter.specific
+package eu.darken.sdmse.systemcleaner.core.filter.stock
 
 import dagger.Binds
 import dagger.Module
@@ -14,47 +14,43 @@ import eu.darken.sdmse.common.ca.CaString
 import eu.darken.sdmse.common.ca.toCaDrawable
 import eu.darken.sdmse.common.ca.toCaString
 import eu.darken.sdmse.common.datastore.value
-import eu.darken.sdmse.common.debug.logging.Logging.Priority.INFO
 import eu.darken.sdmse.common.debug.logging.log
 import eu.darken.sdmse.common.debug.logging.logTag
 import eu.darken.sdmse.common.files.APathLookup
-import eu.darken.sdmse.common.files.segs
-import eu.darken.sdmse.common.root.RootManager
-import eu.darken.sdmse.common.root.canUseRootNow
 import eu.darken.sdmse.systemcleaner.core.BaseSieve
 import eu.darken.sdmse.systemcleaner.core.SystemCleanerSettings
 import eu.darken.sdmse.systemcleaner.core.filter.SystemCleanerFilter
+import java.io.File
 import javax.inject.Inject
 import javax.inject.Provider
 
-class DataLoggerFilter @Inject constructor(
+class LinuxFilesFilter @Inject constructor(
     private val baseSieveFactory: BaseSieve.Factory,
     private val areaManager: DataAreaManager,
 ) : SystemCleanerFilter {
 
-    override suspend fun getIcon(): CaDrawable = R.drawable.ic_baseline_format_list_bulleted_24.toCaDrawable()
+    override suspend fun getIcon(): CaDrawable = R.drawable.ic_os_linux.toCaDrawable()
 
-    override suspend fun getLabel(): CaString = R.string.systemcleaner_filter_datalogger_label.toCaString()
+    override suspend fun getLabel(): CaString = R.string.systemcleaner_filter_linuxfiles_label.toCaString()
 
-    override suspend fun getDescription(): CaString = R.string.systemcleaner_filter_datalogger_summary.toCaString()
+    override suspend fun getDescription(): CaString = R.string.systemcleaner_filter_linuxfiles_summary.toCaString()
 
     override suspend fun targetAreas(): Set<DataArea.Type> = setOf(
-        DataArea.Type.DATA,
+        DataArea.Type.SDCARD,
     )
 
     private lateinit var sieve: BaseSieve
 
     override suspend fun initialize() {
-        val config = BaseSieve.Config(
-            areaTypes = targetAreas(),
-            targetTypes = setOf(BaseSieve.TargetType.FILE),
-            pathAncestors = setOf(
-                segs("logger"),
-                segs("log"),
-                segs("log_other_mode"),
-            ),
+        val regexes = setOf(
+            Regex("^(?:[\\W\\w]+/\\.Trash)$".replace("/", "\\" + File.separator)),
+            Regex("^(?:[\\W\\w]+/\\.Trash-[0-9]{1,4})$".replace("/", "\\" + File.separator)),
         )
-
+        val config = BaseSieve.Config(
+            targetTypes = setOf(BaseSieve.TargetType.DIRECTORY),
+            areaTypes = targetAreas(),
+            regexes = regexes,
+        )
         sieve = baseSieveFactory.create(config)
         log(TAG) { "initialized()" }
     }
@@ -69,17 +65,9 @@ class DataLoggerFilter @Inject constructor(
     @Reusable
     class Factory @Inject constructor(
         private val settings: SystemCleanerSettings,
-        private val filterProvider: Provider<DataLoggerFilter>,
-        private val rootManager: RootManager,
+        private val filterProvider: Provider<LinuxFilesFilter>
     ) : SystemCleanerFilter.Factory {
-
-        override suspend fun isEnabled(): Boolean {
-            val enabled = settings.filterDataLoggerEnabled.value()
-            val useRoot = rootManager.canUseRootNow()
-            if (enabled && !useRoot) log(TAG, INFO) { "Filter is enabled, but requires root, which is unavailable." }
-            return enabled && useRoot
-        }
-
+        override suspend fun isEnabled(): Boolean = settings.filterLinuxFilesEnabled.value()
         override suspend fun create(): SystemCleanerFilter = filterProvider.get()
     }
 
@@ -90,6 +78,6 @@ class DataLoggerFilter @Inject constructor(
     }
 
     companion object {
-        private val TAG = logTag("SystemCleaner", "Filter", "DataLogger")
+        private val TAG = logTag("SystemCleaner", "Filter", "LinuxFiles")
     }
 }
