+++ app/src/test/java/eu/darken/sdmse/systemcleaner/core/filter/generic/AdvertisementFilterTest.kt
@@ -3,6 +3,7 @@ package eu.darken.sdmse.systemcleaner.core.filter.generic
 import eu.darken.sdmse.common.areas.DataArea.Type.SDCARD
 import eu.darken.sdmse.systemcleaner.core.BaseSieve
 import eu.darken.sdmse.systemcleaner.core.filter.SystemCleanerFilterTest
+import eu.darken.sdmse.systemcleaner.core.filter.stock.AdvertisementFilter
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
