+++ app/src/test/java/eu/darken/sdmse/systemcleaner/core/filter/generic/AnalyticsFilterTest.kt
@@ -4,6 +4,7 @@ import eu.darken.sdmse.common.areas.DataArea.Type
 import eu.darken.sdmse.common.areas.currentAreas
 import eu.darken.sdmse.systemcleaner.core.BaseSieve
 import eu.darken.sdmse.systemcleaner.core.filter.SystemCleanerFilterTest
+import eu.darken.sdmse.systemcleaner.core.filter.stock.AnalyticsFilter
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
