+++ app/src/test/java/eu/darken/sdmse/systemcleaner/core/filter/generic/TempFilesFilterTest.kt
@@ -4,6 +4,7 @@ import eu.darken.sdmse.common.areas.currentAreas
 import eu.darken.sdmse.common.rngString
 import eu.darken.sdmse.systemcleaner.core.BaseSieve
 import eu.darken.sdmse.systemcleaner.core.filter.SystemCleanerFilterTest
+import eu.darken.sdmse.systemcleaner.core.filter.stock.TempFilesFilter
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
